Create SpatialDataViewer (#120)
* Add elemtents to layer enum

* toggle visibility upon changing cs

* Add metadata visibililty state

* Toggle active_in_cs

* Handle presence in multiple cs

* Add landmark registration case

* Refactor test and utils

* docstring adjustment


* Use qtpy

* small refactor

* Remove SpatialDataLayers

* Fix after merge

* Active_in_cs private

* private meta last

* refactor interactive to _sdatawidgets

* Add current_cs metadata

* add spatialdata viewer

* Viewer refactor

* active layer check

* Fix camera issue

* remove self

* Refactor

* Refactor viewer reference

* Refactor

* Add keybind and public function

* Add test

* Use id

* test additional metadata

* fix test

* remove invalid test

* Revert id

* remove unnecessary shutil

* add None metadata

* Remove shapes_type and add refactor to region_key

* Refactor to region_key

---------
